+++ src/games/super-mario-land/game.ts
@@ -28,7 +28,7 @@ const config: GameConfig = {
   physics: {
     default: "arcade",
     arcade: {
-      gravity: { y: 0 },
+      gravity: { y: 5000 },
       debug: false
     }
   },