+++ src/games/super-mario-land/scenes/boot-scene.ts
@@ -5,7 +5,13 @@
  * @license      Digitsensitive
  */
 
+import { AnimationHelper } from "../helpers/animation-helper";
+
 export class BootScene extends Phaser.Scene {
+  // helpers
+  private animationHelperInstance: AnimationHelper;
+
+  // graphics
   private loadingBar: Phaser.GameObjects.Graphics;
   private progressBar: Phaser.GameObjects.Graphics;
 
@@ -40,6 +46,10 @@ export class BootScene extends Phaser.Scene {
     this.load.on(
       "complete",
       function() {
+        this.animationHelperInstance = new AnimationHelper(
+          this,
+          this.cache.json.get("animationJSON")
+        );
         this.progressBar.destroy();
         this.loadingBar.destroy();
       },